app/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==7.0.1
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==6.0.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
whitenoise==5.3.0
